+++ src/Filterable.php
@@ -2,25 +2,29 @@
 
 namespace Vixen\Enums;
 
-use Illuminate\Support\Collection;
-
 trait Filterable
 {
     /**
      * @param array<string> $cases
      */
-    public static function only(array $cases): Collection
+    public static function only(array $cases): EnumCollection
     {
-        return collect(self::cases())
-            ->filter(fn ($case) => in_array($case->name, $cases));
+        return new EnumCollection(
+            collect(self::cases())
+                ->filter(fn ($case) => in_array($case->name, $cases))
+                ->values()
+        );
     }
 
     /**
      * @param array<string> $cases
      */
-    public static function except(array $cases): Collection
+    public static function except(array $cases): EnumCollection
     {
-        return collect(self::cases())
-            ->filter(fn ($case) => !in_array($case->name, $cases));
+        return new EnumCollection(
+            collect(self::cases())
+                ->filter(fn ($case) => !in_array($case->name, $cases))
+                ->values()
+        );
     }
 }