+++ tests/Samples/HasDescriptionsEnum.php
@@ -4,12 +4,14 @@ namespace Vixen\Enums\Tests\Samples;
 
 use Vixen\Enums\Attributes\Description;
 use Vixen\Enums\Attributes\Label;
+use Vixen\Enums\Filterable;
 use Vixen\Enums\HasDescriptions;
 use Vixen\Enums\HasLabels;
 use Vixen\Enums\HasOptions;
 
 enum HasDescriptionsEnum: string
 {
+    use Filterable;
     use HasDescriptions;
     use HasLabels;
     use HasOptions;